Editing all the endpoints for my models

+++ controllers/wishlistController.js
@@ -1,7 +1,21 @@
 const wishlistService=require('../services/wishListService')
+const db=require('../models')
+
 
 const getAllWishedList=async(req,res)=>{
+
     try{
+        const student=await db.Student.findByPk(req.params.studentId)
+        if(!student){
+            return res.status(401).json({message:"User not found!"})
+        }
+        const residence=await db.Residence.findByPk(req.params.residenceId)
+        if(!residence){
+            return res.status(401).json({message:"Resiedence Is not found!"})
+        }
+
+
+
         const wish=await wishlistService.getWishedList()
         if (wish.length===0){
             return res.status(201).json({message:"No liked items to show"})
@@ -15,6 +29,16 @@ const getAllWishedList=async(req,res)=>{
 
 const addToWishlist=async(req,res)=>{
     try{
+        const student=await db.Student.findByPk(req.params.studentId)
+        if(!user){
+            return res.status(401).json({message:"User not found!"})
+        }
+        const residence=await db.Residence.findByPk(req.params.residenceId)
+        if(!residence){
+            return res.status(401).json({message:"Resiedence Is not found!"})
+        }
+
+
         const wish=await wishlistService.addToWishList(req.body)
         res.status(201).json({message:"resident liked"})
     }
@@ -26,6 +50,16 @@ const addToWishlist=async(req,res)=>{
 
 const removeFromWishlist=async(req,res)=>{
     try{
+        const student=await db.Student.findByPk(req.params.studentId)
+        if(!user){
+            return res.status(401).json({message:"User not found!"})
+        }
+        const residence=await db.Residence.findByPk(req.params.residenceId)
+        if(!residence){
+            return res.status(401).json({message:"Resiedence Is not found!"})
+        }
+
+
         const wish=await wishlistService.removeFromWishList(req.params)
         if(wish===null){
             return res.status(404).json({message:"didn't remove ir"})
